+++ packages/mini_program_vscode/src/statusTree.ts
@@ -1,6 +1,7 @@
 import * as vscode from 'vscode';
 
 import {
+  FirebaseAuthStatus,
   FirebaseHostEndpointStatus,
   StatusTreeSection,
   StatusTreeRow,
@@ -30,6 +31,7 @@ export class MiniProgramStatusTreeProvider implements vscode.TreeDataProvider<St
   private errorMessage: string | undefined;
   private report: WorkflowStatusReport | undefined;
   private firebaseHostEndpoint: FirebaseHostEndpointStatus | undefined;
+  private firebaseAuthStatus: FirebaseAuthStatus | undefined;
 
   readonly onDidChangeTreeData = this.changeEmitter.event;
 
@@ -38,6 +40,7 @@ export class MiniProgramStatusTreeProvider implements vscode.TreeDataProvider<St
     this.report = report;
     this.sections = buildStatusTreeSections(report, {
       firebaseHostEndpoint: this.firebaseHostEndpoint,
+      firebaseAuthStatus: this.firebaseAuthStatus,
     });
     this.changeEmitter.fire(undefined);
   }
@@ -47,6 +50,17 @@ export class MiniProgramStatusTreeProvider implements vscode.TreeDataProvider<St
     this.firebaseHostEndpoint = status;
     this.sections = buildStatusTreeSections(this.report, {
       firebaseHostEndpoint: this.firebaseHostEndpoint,
+      firebaseAuthStatus: this.firebaseAuthStatus,
+    });
+    this.changeEmitter.fire(undefined);
+  }
+
+  setFirebaseAuthStatus(status: FirebaseAuthStatus): void {
+    this.errorMessage = undefined;
+    this.firebaseAuthStatus = status;
+    this.sections = buildStatusTreeSections(this.report, {
+      firebaseHostEndpoint: this.firebaseHostEndpoint,
+      firebaseAuthStatus: this.firebaseAuthStatus,
     });
     this.changeEmitter.fire(undefined);
   }